multimodal-rag-groq/requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
langchain-community
faiss-cpu
sentence-transformers
transformers
torch
accelerate
pillow
unstructured
unstructured[pdf]
pdfminer.six
opencv-python
